content_module: geocoding: improve result bounding
Sort the result against center is a sub-optimal solution. This does not
work when there are multiple results and the user wants result that is
far from the center (e.g. Taichung Train Station can not be found if we
sent senter to Taipei). To solve this, we use multiple bounding box to
bound region of interest, and only return results that are in the
bouding boxs. Since Google Geocoding API only support passing one
bounding box, we pass the max bouding box that cover all bounding box
requested by the user.