engine: add input transformation function
Summary:
Allow user input to be first transform by input transformation set the
previous content module. This is used to solve the case when doesn't
press the postback button but input the title of postback button ead.
This also solves the problem if platforms that does not have postback on.

Test Plan: make

Reviewers: yunchiao.li

Differential Revision: https://compose.phacility.com/D38